requirements.txt
#numpy==1.19

torch
torchvision
torchaudio
scipy
#scikit-learn==0.19.2
opencv-python
pytorch-argus==1.0.0
numpy==1.25.2
scipy==1.11.1
opencv-python-headless==4.8.0.74
pandas==2.0.3
scikit-learn==1.3.0
scikit-image==0.21.0
matplotlib==3.7.2
notebook==7.0.1
hydra-core==1.3.2
optuna==2.10.1  # Older version as newer are not compatible with sweeper
hydra-optuna-sweeper==1.2.0
mlflow-skinny==2.5.0
tqdm==4.65.0
mypy==1.4.1